feat(compiler): complete Stage 152 (TD-OPTION-AND-THEN-I32-MISMATCH) – fix AdtLayout Param substitution for generic enums
Changes:
  - src/codegen/mir_translation/layouts.rs: revert I64 fallback for Param in enum payloads; use standard mir_type_to_emit_type_with_layouts_and_mono
  - src/codegen/mir_translation/types.rs: add substitute_adt_layout helper; apply substitution in mir_type_to_emit_type_with_layouts_and_mono when Adt substs are concrete (no Param); per §1.0 principles 6/9/10

Updates:
  - Cargo.toml: version to v0.676.0
  - README.md, RELEASE_NOTES.md: update version and stage 152 status (TD-OPTION-AND-THEN-I32-MISMATCH fixed)
  - docs/develop/v0/calibration-data.md: add Stage 152 entry (0 regressions, 0 new tests)
  - docs/develop/v0/tech-debt-register.md: mark TD-OPTION-AND-THEN-I32-MISMATCH as resolved; add TD-CALL-DEST-TYPE-SUBSTS as new P3 follow-up
  - docs/worklog.md: append Stage 152 work log
  - tests/v0/stage40/plan/stage40_1_tests.rs: revert to original i32 versions for Option/Result map/and_then tests

New:
  - docs/develop/v0/stage-152/dev-log.md: complete development log (root cause, revert, substitute_adt_layout, discovered TD-CALL-DEST-TYPE-SUBSTS)

Refs:
  - docs/develop/v0/stage-152/dev-log.md
  - docs/develop/v0/tech-debt-register.md
  - docs/develop/v0/calibration-data.md
  - docs/worklog.md

Version: v0.676.0

+++ tests/v0/stage40/plan/stage40_1_tests.rs
@@ -80,16 +80,15 @@ fn main() -> i32 {
 /// Stage 40.1 positive 3: `Option::and_then` on `Some(v)` calls `f(v)`.
 #[test]
 fn stage40_1_pos_option_and_then_some() {
-    // Stage 151: Changed from i32 to i64 — i32 Option in prelude and_then
-    // has a type mismatch (prelude and_then returns {i32,i64} but caller
-    // expects {i32,i32}). This is a known prelude limitation (TD-OPTION-AND-THEN-I32-MISMATCH).
+    // Stage 152: Reverted to i32 — Stage 152's substitute_adt_layout fix
+    // correctly handles i32 (no longer needs I64 fallback from Stage 151).
     let code = r#"
-fn half_even(x: i64) -> Option<i64> {
+fn half_even(x: i32) -> Option<i32> {
     if x % 2 == 0 { Option::Some(x / 2) } else { Option::None }
 }
 fn main() -> i32 {
-    let x: Option<i64> = Option::Some(42i64);
-    let y: Option<i64> = x.and_then(half_even);
+    let x: Option<i32> = Option::Some(42i32);
+    let y: Option<i32> = x.and_then(half_even);
     match y {
         Some(v) => { println!("{}", v); 0 }
         None => { println!("none"); 0 },
@@ -121,14 +120,12 @@ fn main() -> i32 {
 /// Stage 40.1 positive 5: `Result::map` on `Ok(v)` returns `Ok(f(v))`.
 #[test]
 fn stage40_1_pos_result_map_ok() {
-    // Stage 151: Changed from i32 to i64 — i32 Result in prelude map
-    // has a type mismatch (prelude map returns {i32,i64} but caller
-    // expects {i32,i32}). This is a known prelude limitation.
+    // Stage 152: Reverted to i32.
     let code = r#"
-fn double(x: i64) -> i64 { x * 2i64 }
+fn double(x: i32) -> i32 { x * 2 }
 fn main() -> i32 {
-    let r: Result<i64, i64> = Result::Ok(21i64);
-    let r2: Result<i64, i64> = r.map(double);
+    let r: Result<i32, i32> = Result::Ok(21i32);
+    let r2: Result<i32, i32> = r.map(double);
     match r2 {
         Ok(v) => { println!("{}", v); 0 }
         Err(_) => { println!("err"); 0 },
@@ -141,14 +138,14 @@ fn main() -> i32 {
 /// Stage 40.1 positive 6: `Result::map` on `Err(e)` propagates `Err(e)`.
 #[test]
 fn stage40_1_pos_result_map_err() {
-    // Stage 151: Changed from i32 to i64.
+    // Stage 152: Reverted to i32.
     let code = r#"
-fn double(x: i64) -> i64 { x * 2i64 }
+fn double(x: i32) -> i32 { x * 2 }
 fn main() -> i32 {
-    let r: Result<i64, i64> = Result::Err(99i64);
-    let r2: Result<i64, i64> = r.map(double);
+    let r: Result<i32, i32> = Result::Err(99i32);
+    let r2: Result<i32, i32> = r.map(double);
     match r2 {
-        Ok(v) => { println!("{}", v); v as i32 }
+        Ok(v) => { println!("{}", v); v }
         Err(e) => { println!("err: {}", e); 0 },
     }
 }
@@ -159,14 +156,14 @@ fn main() -> i32 {
 /// Stage 40.1 positive 7: `Result::and_then` on `Ok(v)` calls `f(v)`.
 #[test]
 fn stage40_1_pos_result_and_then_ok() {
-    // Stage 151: Changed from i32 to i64.
+    // Stage 152: Reverted to i32.
     let code = r#"
-fn half_even(x: i64) -> Result<i64, i64> {
-    if x % 2i64 == 0i64 { Result::Ok(x / 2i64) } else { Result::Err(x) }
+fn half_even(x: i32) -> Result<i32, i32> {
+    if x % 2 == 0 { Result::Ok(x / 2) } else { Result::Err(x) }
 }
 fn main() -> i32 {
-    let r: Result<i64, i64> = Result::Ok(42i64);
-    let r2: Result<i64, i64> = r.and_then(half_even);
+    let r: Result<i32, i32> = Result::Ok(42i32);
+    let r2: Result<i32, i32> = r.and_then(half_even);
     match r2 {
         Ok(v) => { println!("{}", v); 0 }
         Err(e) => { println!("err: {}", e); 0 },
@@ -179,16 +176,16 @@ fn main() -> i32 {
 /// Stage 40.1 positive 8: `Result::and_then` on `Err(e)` propagates `Err(e)`.
 #[test]
 fn stage40_1_pos_result_and_then_err() {
-    // Stage 151: Changed from i32 to i64.
+    // Stage 152: Reverted to i32.
     let code = r#"
-fn half_even(x: i64) -> Result<i64, i64> {
-    if x % 2i64 == 0i64 { Result::Ok(x / 2i64) } else { Result::Err(x) }
+fn half_even(x: i32) -> Result<i32, i32> {
+    if x % 2 == 0 { Result::Ok(x / 2) } else { Result::Err(x) }
 }
 fn main() -> i32 {
-    let r: Result<i64, i64> = Result::Err(99i64);
-    let r2: Result<i64, i64> = r.and_then(half_even);
+    let r: Result<i32, i32> = Result::Err(99i32);
+    let r2: Result<i32, i32> = r.and_then(half_even);
     match r2 {
-        Ok(v) => { println!("{}", v); v as i32 }
+        Ok(v) => { println!("{}", v); v }
         Err(e) => { println!("err: {}", e); 0 },
     }
 }
@@ -414,14 +411,14 @@ fn main() -> i32 {
 #[test]
 fn stage40_1_neg_codegen_ir_validity() {
     use std::process::Command;
-    // Stage 151: Changed from i32 to i64.
+    // Stage 152: Reverted to i32.
     let code = r#"
-fn double(x: i64) -> i64 { x * 2i64 }
+fn double(x: i32) -> i32 { x * 2 }
 fn main() -> i32 {
-    let x: Option<i64> = Option::Some(42i64);
-    let y: Option<i64> = x.map(double);
+    let x: Option<i32> = Option::Some(42i32);
+    let y: Option<i32> = x.map(double);
     match y {
-        Some(v) => v as i32,
+        Some(v) => v,
         None => 0,
     }
 }
